requirements.txt
langchain==0.0.348
langchain_experimental==0.0.45
pandas==2.0.3
python-dotenv==1.0.0
streamlit==1.25.0
openai
